requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-google-genai
langchain-mistralai
langchain-groq
openai
groq
google-generativeai
mistralai
python-dotenv
pydantic
requests
tqdm
faiss-cpu
chromadb
tiktoken
fastapi
uvicorn
pypdf
unstructured
beautifulsoup4
lxml
langchain-huggingface
streamlit
